initialise category class, move related predicates to categories package
