+++ src/main/java/seedu/address/logic/parser/FilterCommandParser.java
@@ -6,7 +6,7 @@ import java.util.Arrays;
 
 import seedu.address.logic.commands.FilterCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.restaurant.CuisineContainsKeywordsPredicate;
+import seedu.address.model.restaurant.categories.CuisineContainsKeywordsPredicate;
 
 /**
  * Parses input arguments and creates new FilterCommand object
