+++ src/main/java/seedu/address/model/restaurant/categories/Category.java
@@ -1,0 +1,13 @@
+package seedu.address.model.restaurant.categories;
+
+import java.util.Optional;
+
+public class Category {
+    private Optional<Cuisine> cuisine;
+    private Optional<Occasion> occasion;
+
+    public Category(Cuisine cuisine, Occasion occasion) {
+        this.cuisine = Optional.ofNullable(cuisine);
+        this.occasion = Optional.ofNullable(occasion);
+    }
+}
